stm32: extract backupdomain into mod
