+++ embassy-stm32/src/rcc/f4.rs
@@ -8,8 +8,8 @@ use crate::gpio::sealed::AFType;
 use crate::gpio::Speed;
 use crate::pac::rcc::vals::{Hpre, Ppre, Sw};
 use crate::pac::{FLASH, PWR, RCC};
+use crate::rcc::bd::{BackupDomain, RtcClockSource};
 use crate::rcc::{set_freqs, Clocks};
-use crate::rtc::{Rtc, RtcClockSource};
 use crate::time::Hertz;
 use crate::{peripherals, Peripheral};
 
@@ -470,7 +470,7 @@ pub(crate) unsafe fn init(config: Config) {
     }
 
     config.rtc.map(|clock_source| {
-        Rtc::set_clock_source(clock_source);
+        BackupDomain::set_rtc_clock_source(clock_source);
     });
 
     let rtc = match config.rtc {
